+++ coursescheduler/src/main/java/vn/edu/haui/scheduler/infrastructure/persistence/jdbc/JdbcHocPhanRepository.java
@@ -10,7 +10,7 @@ import java.util.Optional;
 public class JdbcHocPhanRepository implements HocPhanRepositoryPort
 {
 	@Override
-	public Optional<Integer> findIdByMaHocPhan(String maHocPhan) throws SQLException
+	public Optional<Long> findIdByMaHocPhan(String maHocPhan) throws SQLException
 	{
 		if(maHocPhan == null) return Optional.empty();
 		String sql = "SELECT id FROM hoc_phan WHERE ma_hoc_phan = ?";
@@ -18,14 +18,14 @@ public class JdbcHocPhanRepository implements HocPhanRepositoryPort
 				PreparedStatement ps = conn.prepareStatement(sql)) {
 			ps.setString(1, maHocPhan);
 			try (ResultSet rs = ps.executeQuery()) {
-				if(rs.next()) return Optional.of(rs.getInt(1));
+				if(rs.next()) return Optional.of(rs.getLong(1));
 			}
 		}
 		return Optional.empty();
 	}
 
 	@Override
-	public int save(HocPhan hocPhan) throws SQLException
+	public Long save(HocPhan hocPhan) throws SQLException
 	{
 		String insert = "INSERT INTO hoc_phan (ma_hoc_phan, ten_hoc_phan, so_tin_chi) VALUES (?,?,?)";
 		try (Connection conn = DataSourceProvider.getDataSource().getConnection();
@@ -36,7 +36,7 @@ public class JdbcHocPhanRepository implements HocPhanRepositoryPort
 			else ps.setNull(3, Types.INTEGER);
 			ps.executeUpdate();
 			try (ResultSet rs = ps.getGeneratedKeys()) {
-				if(rs.next()) return rs.getInt(1);
+				if(rs.next()) return rs.getLong(1);
 			}
 		}
 		throw new SQLException("Cannot create hoc_phan");
