+++ coursescheduler/src/main/java/vn/edu/haui/scheduler/infrastructure/persistence/jdbc/JdbcTepTaiLenRepository.java
@@ -9,12 +9,12 @@ import java.sql.*;
 public class JdbcTepTaiLenRepository implements TepTaiLenRepositoryPort
 {
 	@Override
-	public int saveMetadata(int nguoiTaoId, File file, String loaiTep) throws SQLException
+	public Long saveMetadata(Long nguoiTaoId, File file, String loaiTep) throws SQLException
 	{
 		String sql = "INSERT INTO tep_tai_len (nguoi_tao_id, ten_tep_goc, loai_tep, duong_dan, storage_type, kich_thuoc) VALUES (?,?,?,?,?,?)";
 		try (Connection conn = DataSourceProvider.getDataSource().getConnection();
 				PreparedStatement ps = conn.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
-			ps.setInt(1, nguoiTaoId);
+			ps.setLong(1, nguoiTaoId);
 			ps.setString(2, file.getName());
 			ps.setString(3, loaiTep);
 			ps.setString(4, file.getAbsolutePath());
@@ -22,7 +22,7 @@ public class JdbcTepTaiLenRepository implements TepTaiLenRepositoryPort
 			ps.setLong(6, file.length());
 			ps.executeUpdate();
 			try (ResultSet rs = ps.getGeneratedKeys()) {
-				if(rs.next()) return rs.getInt(1);
+				if(rs.next()) return rs.getLong(1);
 			}
 		}
 		throw new SQLException("Cannot create tep_tai_len");
